lifted state correctly (i hope)
